requirements.txt
gunicorn==20.1.0
flask==2.2.3
pymongo==4.3.3
boto3==1.26.13
pillow==9.3.0
pytz==2023.3